+++ src/main/java/project/moodipie/config/GlobalExceptionHandler.java
@@ -6,9 +6,9 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import project.moodipie.error.ApiRes;
-import project.moodipie.error.ErrorCode;
-import project.moodipie.error.FieldErrors;
+import project.moodipie.response.ApiRes;
+import project.moodipie.response.error.ErrorCode;
+import project.moodipie.response.error.FieldErrors;
 
 @RestControllerAdvice
 @Slf4j
