+++ src/main/java/project/moodipie/response/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package project.moodipie.error;
+package project.moodipie.response.error;
 
 import lombok.Getter;
